@@ -1,10 +1,17 @@
 module github.com/mantidtech/tplr
 
-go 1.14
+go 1.20
 
 require (
-	github.com/mantidtech/wordcase v1.0.0
-	github.com/stretchr/testify v1.5.1
-	golang.org/x/sys v0.1.0
-	gopkg.in/yaml.v2 v2.3.0
+	github.com/mantidtech/wordcase v1.0.1
+	github.com/stretchr/testify v1.8.2
+	golang.org/x/exp v0.0.0-20230519143937-03e91628a987
+	golang.org/x/sys v0.7.0
+	gopkg.in/yaml.v2 v2.4.0
+)
+
+require (
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
