@@ -10,7 +10,7 @@ import (
 // TestLogicFunctions provides unit test coverage for LogicFunctions
 func TestLogicFunctions(t *testing.T) {
 	fn := Functions()
-	assert.Len(t, fn, 5, "weakly ensuring functions haven't been added/removed without updating tests")
+	assert.Len(t, fn, 3, "weakly ensuring functions haven't been added/removed without updating tests")
 }
 
 // TestWhen provides unit test coverage for When()
@@ -253,154 +253,6 @@ func TestIsZero(t *testing.T) {
 				"value": 2,
 			},
 			Want: "two",
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.Name, helper.TemplateTest(tt, Functions()))
-	}
-}
-
-// TestAnd provides unit test coverage for And()
-func TestAnd(t *testing.T) {
-	tests := []helper.TestSet{
-		{
-			Name:     "no args",
-			Template: `{{ and }}`,
-			Args:     helper.TestArgs{},
-			Want:     "",
-		},
-		{
-			Name:     "bool false",
-			Template: `{{ and .A }}`,
-			Args: helper.TestArgs{
-				"A": false,
-			},
-			Want: "",
-		},
-		{
-			Name:     "bool true",
-			Template: `{{ and .A }}`,
-			Args: helper.TestArgs{
-				"A": true,
-			},
-			Want: "true",
-		},
-		{
-			Name:     "int 9",
-			Template: `{{ and .A }}`,
-			Args: helper.TestArgs{
-				"A": 9,
-			},
-			Want: "9",
-		},
-		{
-			Name:     "2 empty args",
-			Template: `{{ and .A .B }}`,
-			Args: helper.TestArgs{
-				"A": "",
-				"B": 0,
-			},
-			Want: "",
-		},
-		{
-			Name:     "3 true args",
-			Template: `{{ and .A .B .C }}`,
-			Args: helper.TestArgs{
-				"A": 2,
-				"B": 3,
-				"C": "X",
-			},
-			Want: "X",
-		},
-		{
-			Name:     "3 args, 1 false",
-			Template: `{{ and .A .B .C }}`,
-			Args: helper.TestArgs{
-				"A": 2,
-				"B": 0,
-				"C": "X",
-			},
-			Want: "",
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.Name, helper.TemplateTest(tt, Functions()))
-	}
-}
-
-// TestOr provides unit test coverage for Or()
-func TestOr(t *testing.T) {
-	tests := []helper.TestSet{
-		{
-			Name:     "no args",
-			Template: `{{ or }}`,
-			Args:     helper.TestArgs{},
-			Want:     "",
-		},
-		{
-			Name:     "bool false",
-			Template: `{{ or .A }}`,
-			Args: helper.TestArgs{
-				"A": false,
-			},
-			Want: "",
-		},
-		{
-			Name:     "bool true",
-			Template: `{{ or .A }}`,
-			Args: helper.TestArgs{
-				"A": true,
-			},
-			Want: "true",
-		},
-		{
-			Name:     "int 9",
-			Template: `{{ or .A }}`,
-			Args: helper.TestArgs{
-				"A": 9,
-			},
-			Want: "9",
-		},
-		{
-			Name:     "2 empty args",
-			Template: `{{ or .A .B }}`,
-			Args: helper.TestArgs{
-				"A": "",
-				"B": 0,
-			},
-			Want: "",
-		},
-		{
-			Name:     "3 true args",
-			Template: `{{ or .A .B .C }}`,
-			Args: helper.TestArgs{
-				"A": 2,
-				"B": 3,
-				"C": "X",
-			},
-			Want: "2",
-		},
-		{
-			Name:     "3 args, first false",
-			Template: `{{ or .A .B .C }}`,
-			Args: helper.TestArgs{
-				"A": 0,
-				"B": 2,
-				"C": "X",
-			},
-			Want: "2",
-		},
-		{
-			Name:     "3 args, middle false",
-			Template: `{{ or .A .B .C }}`,
-			Args: helper.TestArgs{
-				"A": 2,
-				"B": 0,
-				"C": "X",
-			},
-			Want: "2",
 		},
 	}
 
